python/pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["lhuang19 <70118730+lhuang19@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio-tools = "^1.57.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve="test-server:serve"
